+++ cmd/stado/plugin.go
@@ -26,6 +26,6 @@ func init() {
 		"Path to the plugin wasm binary (default: <manifest-dir>/plugin.wasm)")
 	pluginCmd.AddCommand(pluginTrustCmd, pluginUntrustCmd, pluginListCmd, pluginInstalledCmd, pluginVerifyCmd,
 		pluginDigestCmd, pluginInstallCmd, pluginRunCmd, pluginGenKeyCmd, pluginSignCmd,
-		pluginGCCmd)
+		pluginGCCmd, pluginDoctorCmd)
 	rootCmd.AddCommand(pluginCmd)
 }